(big) Split stuff up into different files; mv into `./src/`
