+++ src/SplitStr.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include <vector>
+#include <string>
+
+std::vector<std::string> splitStr(std::string str = "", const char delim = ' ');
