+++ src/go
@@ -1,0 +1,7 @@
+#!/usr/bin/env sh
+
+PROGNAME="${1:-"xstate"}"
+
+#g++ -std=c++2a ./"$PROGNAME".cpp -o "$PROGNAME".out && ./"$PROGNAME".out
+
+g++ -std=c++17 *.cpp -o "$PROGNAME".out && ./"$PROGNAME".out
