+++ src/interpreter.cpp
@@ -1,78 +1,6 @@
-// xstate.cpp
-
-/**
- * Originally by Kipras Melnikovas (https://kipras.org) <kipras@kipras.org>
- * MIT Licensed
- *
- *
- * This file (`.cpp`) is not actually needed - everything's inside `xstate.h`,
- * and this is just an example usage.
- *
- *
- * compile with:
- *
- * ```sh
- * g++ -std=c++17 ./xstate.cpp -o xstate.out
- * ```
- *
-*/
-
-#include "xstate.h"
-
-#include <vector>
-#include <string>
-#include <sstream>
-
-std::vector<std::string> splitStr(std::string str = "", const char delim = ' ') {
-    std::string buf;                 // Have a buffer string
-    std::stringstream ss(str);       // Insert the string into a stream
-
-    std::vector<std::string> tokens; // Create vector to hold our words
-
-    while (getline(ss, buf, delim)) {
-        tokens.push_back(buf);
-	}
-
-	return tokens;
-}
+#include "Interpreter.h"
 
 namespace xs {
-
-const char *StateMachine::transition(const char *currentState, const char *event)
-{
-	std::vector<std::string> stateAccessTokens = splitStr(currentState, '.');
-
-	// printf("num of tokens %d\n", stateAccessTokens.size());
-
-	// for (auto &token : stateAccessTokens) {
-	// 	printf("token %s\n", token.c_str());
-	// }
-
-	StateMachine *head = this;
-
-	/** go deeper & traverse */
-	for (size_t i = 0; i + 1 < stateAccessTokens.size(); ++i) {
-		const char *stateToken = stateAccessTokens[i].c_str();
-		head = &head->states[stateToken].nested;
-	}
-
-	const int lastIndex = stateAccessTokens.size() - 1;
-	std::string lastStateAccessToken = stateAccessTokens[lastIndex];
-
-	/**
-	 * HALP FIXME
-	 *
-	 * lmfao I spent way too much time here
-	 *
-	 * don't use `const char *` for indexing for sure.
-	 *
-	*/
-	const char *nextState = head->states[lastStateAccessToken].on[event];
-
-	// printf("nextState %s\n", nextState);
-
-	return nextState;
-}
 
 Interpreter::Interpreter(StateMachine *stateMachine)
 	:
@@ -200,72 +128,4 @@ Interpreter *interpret(StateMachine *stateMachine) {
 	return interpreter;
 }
 
-} // namespace xs
-
-// int main() {
-// 	xs::StateMachine machine = {
-// 		.id = "light",
-// 		.initial = "green",
-// 		.states = {
-// 			{
-// 				"green" ,
-// 				{ .on = { { "TIMER", "yellow" } } }
-// 			},
-// 			{
-// 				"yellow",
-// 				{ .on = { { "TIMER", "red"    } } }
-// 			},
-// 			{
-// 				"red"  ,
-// 				{
-// 					.on = { { "TIMER", "red.red-100"  } },
-// 					.nested = {
-// 						.id = "red-brightness",
-// 						.initial = "red-100",
-// 						.states = {
-// 							{
-// 								"red-100",
-// 								{ .on = { { "TIMER", "red.red-0" } } }
-// 							},
-// 							{
-// 								"red-0",
-// 								{ .on = { { "TIMER", "green" } } }
-// 							}
-// 						}
-// 					}
-// 				},
-// 			}
-// 		}
-// 	};
-
-// 	xs::Interpreter *toggleMachine = xs::interpret(machine)
-// 		->logInfo()
-// 		->onStart([]() {
-// 			printf("let's go!\n");
-// 		})
-// 		->onTransition([](xs::Interpreter *self) {
-// 			self->logInfo();
-// 		})
-// 		->onStop([]() {
-// 			printf("oh no we stopped c:\n");
-// 		})
-// 		->start();
-
-// 	toggleMachine->send("TIMER");
-
-// 	toggleMachine->send("TIMER");
-
-// 	toggleMachine->send("TIMER");
-
-// 	toggleMachine->send("TIMER");
-
-// 	toggleMachine->send("TIMER");
-
-// 	toggleMachine->send("TIMER");
-
-// 	toggleMachine->stop();
-
-// 	delete toggleMachine;
-
-// 	return 0;
-// }
+}; // namespace xs
