+++ src/xstate.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include "States.h"
+#include "StateMachine.h"
+#include "State.h"
+#include "Interpreter.h"
